+++ src/common/enums/dom/adjacent-html-place.enum.js
@@ -1,0 +1,8 @@
+const AdjacentHTMLPlace = {
+  BEFORE_BEGIN: `beforebegin`,
+  AFTER_BEGIN: `afterbegin`,
+  BEFORE_END: `beforeend`,
+  AFTER_END: `afterend`,
+};
+
+export { AdjacentHTMLPlace };
