+++ src/helpers/template/index.js
@@ -1,0 +1,1 @@
+export * from "./render-template/render-template";
