+++ src/helpers/template/render-template/render-template.js
@@ -1,0 +1,5 @@
+const renderTemplate = (container, template, place) => {
+  container.insertAdjacentHTML(place, template);
+};
+
+export { renderTemplate };
